requirements.txt
pandas~=2.2.2
streamlit~=1.38.0
toml~=0.10.2
altair~=5.4.1
numpy~=1.26.4
pillow~=10.4.0
scikit-learn~=1.5.0
xgboost~=1.5.0
shap~=0.41.0